requirements.txt
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
cffi==1.11.5
constantly==15.1.0
cryptography==2.4.2
cssselect==1.0.3
Django==2.1.5
django-background-tasks==1.2.0
django-compat==1.0.15
factory-boy==2.11.1
Faker==1.0.1
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
lxml==4.3.0
parsel==1.5.1
psycopg2==2.7.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
python-dateutil==2.7.5
pytz==2018.9
queuelib==1.5.0
Scrapy==1.5.1
service-identity==18.1.0
six==1.12.0
text-unidecode==1.2
Twisted==18.9.0
w3lib==1.20.0
zope.interface==4.6.0
